refactor: recordPicker
