+++ packages/core/client/src/flow/models/data-blocks/table/TableColumnModel.tsx
@@ -220,7 +220,6 @@ TableColumnModel.registerFlow({
         title: ctx.model.collectionField?.title,
       }),
       handler(ctx, params) {
-        console.log('ctx.collectionField', ctx.collectionField);
         const title = ctx.t(params.title || ctx.model.collectionField?.title);
         ctx.model.setProps('title', title);
       },
