+++ packages/core/client/src/flow/models/fields/AssociationFieldModel/index.tsx
@@ -11,3 +11,4 @@ export * from './AssociationFieldModel';
 export * from './RemoteSelectFieldModel';
 export * from './TableAssociationFieldModel';
 export * from './SubFormFieldModel';
+export * from './RecordPickerFieldModel';
